+++ authorization-server/src/main/java/com/server/authorization/config/AuthorizationServerConfig.java
@@ -1,36 +1,21 @@
 package com.server.authorization.config;
 
-import com.nimbusds.jose.jwk.JWKSet;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.proc.SecurityContext;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.interfaces.RSAPrivateKey;
-import java.security.interfaces.RSAPublicKey;
 import java.time.Duration;
-import java.time.Instant;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.UUID;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
 import org.springframework.security.oauth2.core.OAuth2Token;
 import org.springframework.security.oauth2.jose.jws.SignatureAlgorithm;
-import org.springframework.security.oauth2.jwt.JwsHeader;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.jwt.JwtClaimsSet;
 import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
-import org.springframework.security.oauth2.server.authorization.OAuth2TokenType;
 import org.springframework.security.oauth2.server.authorization.client.InMemoryRegisteredClientRepository;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClient;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClientRepository;
@@ -39,23 +24,21 @@ import org.springframework.security.oauth2.server.authorization.config.annotatio
 import org.springframework.security.oauth2.server.authorization.settings.AuthorizationServerSettings;
 import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
 import org.springframework.security.oauth2.server.authorization.settings.TokenSettings;
-import org.springframework.security.oauth2.server.authorization.token.DelegatingOAuth2TokenGenerator;
-import org.springframework.security.oauth2.server.authorization.token.JwtEncodingContext;
-import org.springframework.security.oauth2.server.authorization.token.JwtGenerator;
-import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenCustomizer;
 import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenGenerator;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.util.StringUtils;
 
 @Configuration
 public class AuthorizationServerConfig {
 
   @Bean
   @Order(Ordered.HIGHEST_PRECEDENCE)
-  public SecurityFilterChain authorizationServerSecurityFilterChain(HttpSecurity http) throws Exception {
+  public SecurityFilterChain authorizationServerSecurityFilterChain(
+      HttpSecurity http,
+      OAuth2TokenGenerator<OAuth2Token> jwtTokenGenerator
+  ) throws Exception {
     OAuth2AuthorizationServerConfiguration.applyDefaultSecurity(http);
     http.getConfigurer(OAuth2AuthorizationServerConfigurer.class)
-        .tokenGenerator(jwtTokenGenerator());
+        .tokenGenerator(jwtTokenGenerator);
 
 //    http
 //        // Accept access tokens for User Info and/or Client Registration
@@ -63,45 +46,6 @@ public class AuthorizationServerConfig {
 //            .jwt(Customizer.withDefaults()));
 
     return http.build();
-  }
-
-  @Bean
-  public OAuth2TokenGenerator<OAuth2Token> jwtTokenGenerator() {
-    JwtEncoder jwtEncoder = new NimbusJwtEncoder(jwkSource());
-    JwtGenerator jwtGenerator = new JwtGenerator(jwtEncoder);
-    jwtGenerator.setJwtCustomizer(jwtTokenCustomizer());
-    return new DelegatingOAuth2TokenGenerator(jwtGenerator);
-  }
-
-  @Bean
-  public OAuth2TokenCustomizer<JwtEncodingContext> jwtTokenCustomizer() {
-    return (context) -> {
-      if (OAuth2TokenType.ACCESS_TOKEN.equals(context.getTokenType())) {
-        String issuer;
-        if (context.getAuthorizationServerContext() != null) {
-          issuer = context.getAuthorizationServerContext().getIssuer();
-        } else {
-          issuer = null;
-        }
-
-        JwtClaimsSet.Builder claimsBuilder = context.getClaims();
-        claimsBuilder.claims(claims -> {
-
-        });
-        if (StringUtils.hasText(issuer)) {
-          claimsBuilder.issuer(issuer);
-        }
-        claimsBuilder
-            .subject(context.getPrincipal().getName())
-            .audience(Collections.unmodifiableList(Arrays.asList("", "")))
-            .issuedAt(Instant.now())
-            .expiresAt(Instant.now().plus(context.getRegisteredClient()
-                .getTokenSettings().getAccessTokenTimeToLive()))
-            .id(UUID.randomUUID().toString());
-
-        claimsBuilder.build();
-      }
-    };
   }
 
   @Bean
@@ -123,31 +67,6 @@ public class AuthorizationServerConfig {
         .build();
 
     return new InMemoryRegisteredClientRepository(registeredClient);
-  }
-
-  @Bean
-  public JWKSource<SecurityContext> jwkSource() {
-    KeyPair keyPair = generateRsaKey();
-    RSAPublicKey publicKey = (RSAPublicKey) keyPair.getPublic();
-    RSAPrivateKey privateKey = (RSAPrivateKey) keyPair.getPrivate();
-    RSAKey rsaKey = new RSAKey.Builder(publicKey)
-        .privateKey(privateKey)
-        .keyID(UUID.randomUUID().toString())
-        .build();
-    JWKSet jwkSet = new JWKSet(rsaKey);
-    return new ImmutableJWKSet<>(jwkSet);
-  }
-
-  private static KeyPair generateRsaKey() {
-    KeyPair keyPair;
-    try {
-      KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
-      keyPairGenerator.initialize(2048);
-      keyPair = keyPairGenerator.generateKeyPair();
-    } catch (Exception ex) {
-      throw new IllegalStateException(ex);
-    }
-    return keyPair;
   }
 
   @Bean
