+++ authorization-server/src/main/java/com/server/authorization/config/JwtTokenGeneratorConfig.java
@@ -1,0 +1,90 @@
+package com.server.authorization.config;
+
+import com.nimbusds.jose.jwk.JWKSet;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
+import com.nimbusds.jose.jwk.source.JWKSource;
+import com.nimbusds.jose.proc.SecurityContext;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.interfaces.RSAPrivateKey;
+import java.security.interfaces.RSAPublicKey;
+import java.util.UUID;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.oauth2.core.OAuth2Token;
+import org.springframework.security.oauth2.jwt.JwtClaimsSet;
+import org.springframework.security.oauth2.jwt.JwtEncoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
+import org.springframework.security.oauth2.server.authorization.OAuth2TokenType;
+import org.springframework.security.oauth2.server.authorization.token.DelegatingOAuth2TokenGenerator;
+import org.springframework.security.oauth2.server.authorization.token.JwtEncodingContext;
+import org.springframework.security.oauth2.server.authorization.token.JwtGenerator;
+import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenCustomizer;
+import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenGenerator;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+
+@Configuration
+public class JwtTokenGeneratorConfig {
+
+  @Bean
+  UserDetailsService users() {
+    UserDetails user = User.builder()
+        .username("user")
+        .password("{noop}1234")
+        .roles("USER")
+        .build();
+
+    return new InMemoryUserDetailsManager(user);
+  }
+
+  @Bean
+  public OAuth2TokenGenerator<OAuth2Token> jwtTokenGenerator() {
+    JwtEncoder jwtEncoder = new NimbusJwtEncoder(jwkSource());
+    JwtGenerator jwtGenerator = new JwtGenerator(jwtEncoder);
+    jwtGenerator.setJwtCustomizer(jwtTokenCustomizer(users()));
+    return new DelegatingOAuth2TokenGenerator(jwtGenerator);
+  }
+
+  @Bean
+  public JWKSource<SecurityContext> jwkSource() {
+    KeyPair keyPair = generateRsaKey();
+    RSAPublicKey publicKey = (RSAPublicKey) keyPair.getPublic();
+    RSAPrivateKey privateKey = (RSAPrivateKey) keyPair.getPrivate();
+    RSAKey rsaKey = new RSAKey.Builder(publicKey)
+        .privateKey(privateKey)
+        .keyID(UUID.randomUUID().toString())
+        .build();
+    JWKSet jwkSet = new JWKSet(rsaKey);
+    return new ImmutableJWKSet<>(jwkSet);
+  }
+
+  private static KeyPair generateRsaKey() {
+    KeyPair keyPair;
+    try {
+      KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
+      keyPairGenerator.initialize(2048);
+      keyPair = keyPairGenerator.generateKeyPair();
+    } catch (Exception ex) {
+      throw new IllegalStateException(ex);
+    }
+    return keyPair;
+  }
+
+  @Bean
+  public OAuth2TokenCustomizer<JwtEncodingContext> jwtTokenCustomizer(UserDetailsService userDetailsService) {
+    return (context) -> {
+      if (OAuth2TokenType.ACCESS_TOKEN.equals(context.getTokenType())) {
+        JwtClaimsSet.Builder claimsBuilder = context.getClaims();
+        System.out.println(context.getPrincipal().getName());
+        UserDetails userDetails = userDetailsService.loadUserByUsername("user");
+        claimsBuilder.claims(claims -> {
+          claims.put("userDetails", userDetails);
+        });
+      }
+    };
+  }
+}
